requirements.txt
anthropic>=0.25.0
PyGithub>=2.1.0
python-dotenv>=1.0.0
